% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popReconstruct_hyperparameters.R
\name{calculate_beta}
\alias{calculate_beta}
\title{Calculate the beta hyperparameter for popReconstruct variance
parameters.}
\usage{
calculate_beta(abs_deviation, p, alpha, log_transform)
}
\arguments{
\item{abs_deviation}{[\code{numeric(1)}]\cr
Expert opinion about the absolute deviation \eqn{\eta_{\nu}} of the initial
point estimates for a given probability interval \eqn{p_{\nu}}.}

\item{p}{[\code{numeric(1)}]\cr
Central \eqn{p_{\nu}}\% probability interval used in elicitation statement.}

\item{alpha}{[\code{numeric(1)}]\cr
Alpha (shape) parameter of inverse gamma distribution.}

\item{log_transform}{[\code{logical(1)}]\cr
Whether model component is log transformed in the popReconstruct model.}
}
\value{
[\code{numeric(1)}] Beta (scale) parameter of inverse gamma distribution
for the variance of the popReconstruct model component.
}
\description{
Helper function for calculating the beta parameter of an inverse
gamma distribution given absolute deviation from the popReconstruct
elicitation statement. Used for quantifying measurement error in
popReconstruct components.
}
