% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popReconstruct_draws.R
\name{calculate_ccmpp_input_draws}
\alias{calculate_ccmpp_input_draws}
\title{Helper function to predict ccmpp inputs estimates from spline offsets
and initial estimates.}
\usage{
calculate_ccmpp_input_draws(spline_offset_draws, inputs, detailed_settings)
}
\arguments{
\item{spline_offset_draws}{[\code{list()}] of [\code{data.table()}]\cr
Draws of spline offsets as returned by \code{\link[=calculate_spline_offset_draws]{calculate_spline_offset_draws()}}.}

\item{inputs}{[\code{list()}]\cr
[\code{data.table()}] for each ccmpp input. Requires 'srb', 'asfr', 'baseline',
and 'survival'; and migration estimates provided as just 'net_migration' or
both 'immigration' and 'emigration'. See \strong{Section: Inputs} for more
information on each of the required inputs.}

\item{detailed_settings}{[\code{list()}]\cr
Detailed settings for each ccmpp input and 'population'.}
}
\value{
[\code{list()}] of [\code{data.table()}] containing draws of all ccmpp
inputs.
}
\description{
Helper function to predict ccmpp inputs estimates from spline offsets
and initial estimates.
}
