% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popReconstruct_draws.R
\name{calculate_spline_offset_draws}
\alias{calculate_spline_offset_draws}
\title{Helper function to calculate spline offsets for all ccmpp inputs}
\usage{
calculate_spline_offset_draws(
  offset_draws,
  detailed_settings,
  draw_col_name = "draw"
)
}
\arguments{
\item{offset_draws}{[\code{list()}] of [\code{data.table()}]\cr
Draws of directly modeled offset parameters for all ccmpp inputs as
prepared in the \verb{popReconstruct_*_draws()} functions.}

\item{detailed_settings}{[\code{list()}]\cr
Detailed settings for each ccmpp input and 'population'.}

\item{draw_col_name}{[\code{character(1)}]\cr
Name of column in \code{offset_draws} [\code{data.table()}] that corresponds to
draw index}
}
\value{
[\code{list()}] of [\code{data.table()}] containing draws of spline offsets
parameters for all ccmpp inputs.
}
\description{
Helper function to calculate spline offsets for all ccmpp inputs
}
