% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popReconstruct_settings.R
\name{create_detailed_settings}
\alias{create_detailed_settings}
\title{Helper function to create a detailed list of settings specific to
popReconstruct ccmpp inputs and outputs.}
\usage{
create_detailed_settings(settings)
}
\arguments{
\item{settings}{[\code{list()}]\cr
Named list of settings for running the popReconstruct model with. The
required settings are the same as those required for \code{\link[demCore:ccmpp]{demCore::ccmpp()}},
see \strong{Section: Settings} for these. The optional settings specific to
popReconstruct are described in
\strong{Section: Optional popReconstruct Settings}}
}
\value{
[\code{list()}] of detailed settings for each ccmpp input and
'population'.
}
\description{
Helper function to create a detailed list of settings specific to
popReconstruct ccmpp inputs and outputs.
}
\section{Detailed settings}{

\itemize{
\item measure_type: whether the component is either stock (measured at a specific
point in time) or flow (measured over a period of time). This determines
if the component will have a 'year' or 'year_start' and 'year_end'
columns.
\item id_cols: the id columns defining the input and output [\code{data.table()}].
\item years: the 'year_start' or 'year' values the input/output will have.
\item years_knots: the start of each calendar year interval that will be knots
in the B-spline linear basis functions.
\item B_t: the B-spline basis matrix for a linear spline with knots specified at
'years_knots'.
\item sexes: the 'sex' values the input/output will have.
\item ages: the 'age_start' values the input/output will have.
\item age_knots: the start of each age group interval that will be knots in the
B-spline linear basis functions.
\item B_a: the B-spline basis matrix for a linear spline with knots specified at
'ages_knots'.
\item transformation_name: character name of the transformation.
\item transformation: the transformation used when modeling each component.
\item inverse_transformation: the inverse transformation used when modeling each
component.
\item transformation_arguments: additional arguments to pass to the 'transformation'
and 'inverse_transformation' functions.
\item fixed: whether the ccmpp input component is to be estimated or fixed in the
model.
}
}

