% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popReconstruct_settings.R
\name{create_optional_settings}
\alias{create_optional_settings}
\title{Validate or generate the optional settings used in the popReconstruct
model}
\usage{
create_optional_settings(settings, inputs, data = NULL)
}
\arguments{
\item{settings}{[\code{list()}]\cr
Named list of settings for running the popReconstruct model with. The
required settings are the same as those required for \code{\link[demCore:ccmpp]{demCore::ccmpp()}},
see \strong{Section: Settings} for these. The optional settings specific to
popReconstruct are described in
\strong{Section: Optional popReconstruct Settings}}

\item{inputs}{[\code{list()}]\cr
[\code{data.table()}] for each ccmpp input. Requires 'srb', 'asfr', 'baseline';
mortality estimates provided as just 'survival' or two of 'mx', 'ax', and
'qx'; and migration estimates provided as just 'net_migration' or both
'immigration' and 'emigration'. See \strong{Section: Inputs} for more
information on each of the required inputs.}

\item{data}{[\code{list()}]\cr
[\code{data.table()}] for each type of data input for the model (currently
only 'population'). See \strong{Section: Data} for more information.}
}
\value{
Invisibly returns reference to modified \code{settings}.
}
\description{
Adds general settings for the interval size 'int', each year
for which projected population estimates will exist 'years_projections', and
the years in which census data exists 'census_years.

Also adds/validates settings related to specifying the B-spline linear basis
functions. For all the year-specific input components (all except for
'baseline'), adds the 'k_years_{input_name}' setting which defines the start
of each calendar year interval that will be knots in the B-spline linear
basis functions. For all the age-specific input components (all except for
'srb'), adds the 'k_ages_{input_name}' setting which defines the start of
each age-group interval that will be knots in the B-spline linear basis
functions.
}
