% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popMethods.R
\docType{package}
\name{popMethods}
\alias{popMethods}
\title{Methods for Population Estimation}
\description{
popMethods includes functions for 1) assessing the quality
of population data and correcting for common biases and 2) estimating
population size and distribution.
}
