% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popReconstruct_fit.R, R/popReconstruct_draws.R
\name{popReconstruct_fit}
\alias{popReconstruct_fit}
\alias{popReconstruct_posterior_draws}
\alias{popReconstruct_prior_draws}
\alias{popReconstruct_summarize_draws}
\title{Implement the Bayesian hierarchical PopReconstruct model}
\usage{
popReconstruct_fit(
  inputs,
  data,
  hyperparameters,
  settings,
  value_col,
  software,
  ...
)

popReconstruct_posterior_draws(
  fit,
  inputs,
  settings,
  value_col,
  software,
  method_name
)

popReconstruct_prior_draws(
  inputs,
  hyperparameters,
  settings,
  value_col,
  method_name,
  chunk_size = 100
)

popReconstruct_summarize_draws(
  draws,
  summarize_cols = c("chain", "chain_draw", "draw"),
  value_col = "value",
  ...
)
}
\arguments{
\item{inputs}{[\code{list()}]\cr
[\code{data.table()}] for each ccmpp input. Requires 'srb', 'asfr', 'baseline';
mortality estimates provided as just 'survival' or two of 'mx', 'ax', and
'qx'; and migration estimates provided as just 'net_migration' or both
'immigration' and 'emigration'. See \strong{Section: Inputs} for more
information on each of the required inputs.}

\item{data}{[\code{list()}]\cr
[\code{data.table()}] for each type of data input for the model (currently
only 'population'). See \strong{Section: Data} for more information.}

\item{hyperparameters}{[\code{list()}]\cr
List for each model component ('srb', 'asfr', ...) that itself contains a
list specifying the alpha (shape) and beta (scale) hyperparameters for the
variance of each model component.}

\item{settings}{[\code{list()}]\cr
Named list of settings for running the popReconstruct model with. The
required settings are the same as those required for \code{\link[demCore:ccmpp]{demCore::ccmpp()}},
see \strong{Section: Settings} for these. The optional settings specific to
popReconstruct are described in
\strong{Section: Optional popReconstruct Settings}}

\item{value_col}{[\code{character(1)}]\cr
Name of the column containing the value of interest in each of the
\code{inputs}. Default is 'value'.}

\item{software}{[\code{character()}]\cr
Statistical modeling software to fit popReconstruct with. Either 'stan' or
'tmb' are available.}

\item{...}{For \code{\link[=popReconstruct_fit]{popReconstruct_fit()}} additional arguments to pass to
\code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}} if fitting the model with 'stan' or
\code{\link[optimx:optimx]{optimx::optimx()}} if fitting the model with 'tmb'. For
\code{\link[=popReconstruct_summarize_draws]{popReconstruct_summarize_draws()}} additional arguments to pass to
\code{\link[demUtils:summarize_dt]{demUtils::summarize_dt()}}.}

\item{fit}{[\code{stanfit(1)}] or [\code{sdreport(1)}]\cr
Model fit object returned by \code{\link[=popReconstruct_fit]{popReconstruct_fit()}}.}

\item{method_name}{[\code{character(1)}]\cr
Description to assign to column 'method' in each returned data.table.}

\item{chunk_size}{[\code{integer(1)}]\cr
number of draws to sample from the prior at once.}

\item{draws}{[\code{list()}] of [\code{data.table()}]\cr
Draws from the posterior distribution of the popReconstruct model as
returned by \code{\link[=popReconstruct_posterior_draws]{popReconstruct_posterior_draws()}}.}

\item{summarize_cols}{[\code{character()}]\cr
The \code{id_cols} that should be collapsed and to calculate summary statistics
over.}
}
\description{
The popReconstruct model embeds the cohort component method of population
projection (ccmpp) in a Bayesian hierarchical model. It reconciles
ccmpp input preliminary estimates of sex-ratio at birth, fertility,
mortality, migration, and a baseline population with population data while
accounting for measurement error in these initial estimates.

\code{\link[=popReconstruct_fit]{popReconstruct_fit()}} fits the model using \href{https://mc-stan.org/}{Stan} or
\href{https://kaskr.github.io/adcomp/Introduction.html}{TMB}.

\code{\link[=popReconstruct_posterior_draws]{popReconstruct_posterior_draws()}} produces draws from the posterior
distribution.

\code{\link[=popReconstruct_prior_draws]{popReconstruct_prior_draws()}} produces draws from the prior distribution.

\code{\link[=popReconstruct_summarize_draws]{popReconstruct_summarize_draws()}} produces summary statistics of draws from
the popReconstruct model using \code{\link[demUtils:summarize_dt]{demUtils::summarize_dt()}}. The
\code{summarize_cols} parameter should include 'chain', 'chain_draw' and 'draw'
when the model was fit with stan and just 'draw' when the model was fit with
tmb.
}
\section{Optional popReconstruct Settings}{

\itemize{
\item fixed_parameters: [\code{character()}]\cr
Names of \code{\link[demCore:ccmpp]{demCore::ccmpp()}} input components that should be fixed in
the model, corresponds to the names of the \code{input} list. Defaults to
empty if fitting the model for both sexes, if only 'female' then defaults
to 'srb'.
\item n_draws: [\code{character()}]\cr
The number of draws to take from the output posterior distribution. If
fitting the model with stan then the number of draws is dictated by
\code{chains}, \code{iterations}, \verb{burn-in}, \code{thin} arguments to \code{rstan::sampling};
\code{n_draws} can then be used to subsample down.
\item k_years_{input_name}: [\code{integer()}]\cr
The start of each calendar year interval that will be knots in the
B-spline linear basis functions for the year-specific input estimates,
so not relevant for 'baseline'. Defaults to the setting for 'years' which
is equal to the original popReconstruct model without linear B-splines
to reduce year-specific parameters.
\item k_ages_{input_name}: [\code{integer()}]\cr
The start of each age-group interval that will be knots in the B-spline
linear basis functions for the age-specific input estimates, so not
relevant for 'srb'. Defaults to the setting for 'ages' (or
'ages_{input_name}') which is equal to the original popReconstruct model
without linear B-splines to reduce age-specific parameters.
}
}

\section{Data}{

population: [\code{data.table()}]\cr
\itemize{
\item year: [\code{integer()}] year of population data.
\item sex: [\code{character()}] either 'female', 'male', or 'both'. If 'sexes'
\code{setting} is 'female' can only have 'female' input data.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}. Age groups included must not overlap.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] population count, must be greater than zero.
}
}

\section{\code{popReconstruct_fit} Value}{

If fit using 'stan' an object of class \code{stanfit} and if fit using 'tmb' an
object of class \code{sdreport}. Either represents the fitted results that can be
extracted using \code{\link[=popReconstruct_posterior_draws]{popReconstruct_posterior_draws()}}. Stan has helper packages
that can be used to explore the model fit through \code{shinystan}, \code{bayesplot},
etc.
}

\section{\code{popReconstruct_posterior_draws} and \code{popReconstruct_prior_draws} Value}{


\code{\link[=popReconstruct_posterior_draws]{popReconstruct_posterior_draws()}} and \code{\link[=popReconstruct_prior_draws]{popReconstruct_prior_draws()}} return a
named [\code{list()}] of [\code{data.table()}] for draws from the posterior and
prior distribution respectively for each ccmpp input component along with the
associated offset and spline offset parameters. Draws for the 'variance' and
projected 'population' are also included.

The returned [\code{data.table()}]s will include columns related to the draw
number:
\itemize{
\item draw: the draw index.
\item chain: if using stan to fit the model, the chain the draw came from.
\item chain_draw: if using stan to fit the model, the draw index for that chain.
}

population: [\code{data.table()}]\cr
\itemize{
\item year: [\code{integer()}] mid-year for population estimate.
Corresponds to 'years' \code{setting} plus an extra interval.
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item draw: [\code{integer()}] draw index.
\item \code{value_col}: [\code{numeric()}] projected population count estimates.
}

variance: [\code{data.table()}]\cr
\itemize{
\item parameter: [\code{character()}] the ccmpp component the variance term
corresponds to.
\item draw: [\code{integer()}] draw index.
\item \code{value_col}: [\code{numeric()}] estimated variance value..
}

For each ccmpp component ('srb', 'asfr', etc.) there will be three
[\code{data.table()}] outputs included in the returned [\code{list()}]. These will
each have the same columns as described in the \strong{Section: Inputs} but with
additional draw related columns as described above.
\enumerate{
\item The offset parameters representing the values of the piecewise linear
function at the year- and age-specific knots. These are the estimated
deviations from the initial ccmpp input estimates at the exact knots. The
estimated values are in the transformed modeling space (log transformed for
'srb', logit transformed for 'survival', etc.).
\item The spline offset values calculated by multiplying the fixed B-spline
linear basis functions by the offset parameters. These are the estimated
deviations from the initial ccmpp input estimates at all years and ages. The
estimated values are in the transformed modeling space.
\item The actual ccmpp input posterior draws after combining the spline offset
values with the initial ccmpp input estimates and after applying the inverse
of the transformation used to model each component.
}
}

\section{\code{popReconstruct_summarize_draws} Value}{

Returns a named [\code{list()}] of [\code{data.table()}] for summary statistics
of the input \code{draws}. Each \code{\link[=data.table]{data.table()}} will have the \code{id_cols} for each
component (minus the \code{summarize_cols}) plus summary statistic columns.
The summary statistic columns have the same name as each function specified
in \code{summary_fun} and the quantiles are named like 'q_(\code{probs} * 100)'. See
\verb{[demUtils::summarize_dt()}] for more information.
}

\section{Settings}{


\itemize{
\item years: [\code{numeric()}]\cr
The start of each calendar year interval to project in \code{\link[demCore:ccmpp]{ccmpp()}}.
Corresponds to the 'year_start' column in each of the year-specific inputs.
\item sexes: [\code{character()}]\cr
The sexes being projected in \code{\link[demCore:ccmpp]{ccmpp()}}, must be either just
'female', or 'female' and 'male'. Corresponds to the 'sex' column in each
of the sex-specific inputs.
\item ages: [\code{numeric()}]\cr
The ages being projected in \code{\link[demCore:ccmpp]{ccmpp()}}. Corresponds to the
'age_start' column in each of the standard age-specific inputs.
\item ages_mortality: [\code{numeric()}]\cr
The ages for which mortality parameter estimates are available, includes one
extra age group compared to the 'ages' setting. Corresponds to the
'age_start' column in the mortality [\code{data.table()}] input(s).
\item ages_asfr: [\code{numeric()}]\cr
The assumed female reproductive ages, subset of 'ages'. Corresponds to the
'age_start' column in the age-specific fertility rate (asfr)
[\code{data.table()}] input.
}

}

\section{Inputs}{


srb: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item \code{value_col}: [\code{numeric()}] sex-ratio at birth estimates, must be
greater than zero.
}

asfr: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages_asfr' setting.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] annual age-specific fertility rate
estimates, must be greater than zero.
}

baseline: [\code{data.table()}]\cr
\itemize{
\item year: [\code{integer()}] mid-year for population estimate.
Corresponds to 'years' \code{setting}.
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] baseline year population count estimates,
must be greater than zero.
}

survival: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] survivorship ratio estimates, must be
greater than zero and less than one.
}

mx: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] mortality rate estimates, must be greater
than zero.
}

ax: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] average years lived by those dying in the
interval estimates, must be greater than zero and less than the age
interval length.
}

qx: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] probability of death estimates, must be
greater than zero and less than one.
}

net_migration: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] annual net-migration proportion estimates.
}

immigration: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] annual immigration proportion estimates,
must be greater than zero.
}

emigration: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] annual emigration proportion estimates, must
be greater than zero.
}

}

\examples{
# specify settings for this example
settings = list(
  years = seq(1960, 2000, 5),
  sexes = c("female"),
  ages = seq(0, 80, 5),
  ages_mortality = seq(0, 85, 5),
  ages_asfr = seq(15, 45, 5),

  n_draws = 10
)

# hyperparameters for the variance prior distribution, represents measurement error
hyperparameters <- list(asfr = list(alpha = 1, beta = 0.0109),
                        population = list(alpha = 1, beta = 0.0109),
                        survival = list(alpha = 1, beta = 0.0109),
                        net_migration = list(alpha = 1, beta = 0.0436))

fit_stan <- popMethods::popReconstruct_fit(
  inputs = demCore::burkina_faso_initial_estimates,
  data = demCore::burkina_faso_data,
  hyperparameters = hyperparameters,
  settings = settings,
  value_col = "value",
  software = "stan",
  chains = 1,
  iter = 200,
  warmup = 100,
  thin = 2
)

draws_stan <- popMethods::popReconstruct_posterior_draws(
  fit = fit_stan,
  inputs = demCore::burkina_faso_initial_estimates,
  settings = settings,
  value_col = "value",
  software = "stan",
  method_name = "original"
)

summary_stan <- popMethods::popReconstruct_summarize_draws(
  draws = draws_stan
)

}
\references{
Wheldon, Mark C., Adrian E. Raftery, Samuel J. Clark, and Patrick Gerland.
2013. “Reconstructing Past Populations With Uncertainty From Fragmentary
Data.” Journal of the American Statistical Association 108 (501): 96–110.
\url{https://doi.org/10.1080/01621459.2012.737729}.

\href{https://CRAN.R-project.org/package=popReconstruct}{popReconstruct R Package}

Wheldon, Mark C., Adrian E. Raftery, Samuel J. Clark, and Patrick Gerland.
2015. “Bayesian Reconstruction of Two-Sex Populations by Age: Estimating Sex
Ratios at Birth and Sex Ratios of Mortality.” Journal of the Royal
Statistical Society. Series A: Statistics in Society 178 (4): 977–1007.
\url{https://doi.org/10.1111/rssa.12104}.

\href{https://github.com/markalava/Bayesian-Reconstruction}{markalava/Bayesian-Reconstruction github repo}
}
\seealso{
\code{\link[demCore:ccmpp]{demCore::ccmpp()}}

\code{\link[demUtils:summarize_dt]{demUtils::summarize_dt()}}

\code{\link[=rbindlist_dts]{rbindlist_dts()}}

\code{vignette("popReconstruct")}

\code{vignette("popReconstruct_options")}

Other popReconstruct: 
\code{\link{popReconstruct_hyperparameters}()},
\code{\link{rbindlist_dts}()}
}
\concept{popReconstruct}
