% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popReconstruct_draws.R
\name{predict_spline_offset}
\alias{predict_spline_offset}
\title{Helper function to predict spline offset for one ccmpp input component}
\usage{
predict_spline_offset(dt, B_t = NULL, B_a = NULL, years = NULL, ages = NULL)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr
Input offset estimates for the one ccmpp input component for one draw and
sex.}

\item{B_t}{[\code{matrix()}]\cr
B-spline basis matrix for the year dimension. By default is NULL but must
be provided for all ccmpp inputs except 'baseline'.}

\item{B_a}{[\code{matrix()}]\cr
B-spline basis matrix for the age dimension. By default is NULL but must
be provided for all ccmpp inputs except 'srb'.}

\item{years}{[\code{integer()}]\cr
The start of each calendar year interval for the ccmpp input component. By
default is NULL but must be provided for all ccmpp inputs except 'baseline'.}

\item{ages}{[\code{integer()}]\cr
The start of each age group for the ccmpp input component. By default is
NULL but must be provided for all ccmpp inputs except 'srb'.}
}
\value{
[\code{data.table()}] containing draws of spline offsets for one ccmpp
input.
}
\description{
Helper function to predict spline offset for one ccmpp input component
}
