% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popReconstruct_draws.R
\name{rbindlist_dts}
\alias{rbindlist_dts}
\title{Helper function for rbinding each [\code{data.table()}] nested within a
[\code{list()}]}
\usage{
rbindlist_dts(...)
}
\arguments{
\item{...}{multiple [\code{list()}] of many [\code{data.table()}]\cr
as returned by \code{popReconstruct_draws} or \code{popReconstruct_prior_draws}.}
}
\value{
one [\code{list()}] of many [\code{data.table()}]
}
\description{
Can input multiple lists of data.tables that will be combined
into one list of data.tables. data.tables are matched by the name of each
element of each list.
}
\seealso{
Other popReconstruct: 
\code{\link{popReconstruct_fit}()},
\code{\link{popReconstruct_hyperparameters}()}
}
\concept{popReconstruct}
