% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popReconstruct_fit.R
\name{validate_popReconstruct_data}
\alias{validate_popReconstruct_data}
\title{Helper function to validate input data}
\usage{
validate_popReconstruct_data(data, settings, detailed_settings, value_col)
}
\description{
Assert that the input data.tables includes all expected columns,
combinations of id variables, and that the transformed values are finite.
}
