% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popReconstruct_fit.R
\name{validate_popReconstruct_hyperparameters}
\alias{validate_popReconstruct_hyperparameters}
\title{Helper function to validate hyperparameter values}
\usage{
validate_popReconstruct_hyperparameters(hyperparameters, inputs, settings)
}
\description{
Assert that the list of hyperparameters includes all expected
components that are not fixed in the model.
}
