% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popReconstruct_fit.R
\name{validate_popReconstruct_inputs}
\alias{validate_popReconstruct_inputs}
\title{Helper function to validate initial input data.tables}
\usage{
validate_popReconstruct_inputs(inputs, settings, detailed_settings, value_col)
}
\description{
Assert that each of the inputs once transformed is a finite
value greater than negative Infinity and less than positive Infinity.
}
