% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popReconstruct_settings.R
\name{transform_dt}
\alias{transform_dt}
\title{Helper function to transform values in a data.table}
\usage{
transform_dt(dt, value_col, transformation, transformation_arguments)
}
\arguments{
\item{dt}{[\code{data.table(1)}]\cr
The data.table to apply the transformation functions to.}

\item{value_col}{[\code{character(1)}]\cr
The column in \code{dt} to transform.}

\item{transformation}{[\code{list()}]\cr
NULL if no transformation. Otherwise list of transformation function(s) to
apply to \code{value_col}. If the transformation should only be applied to a
subset of \code{dt} name each function in \code{transformation} with a string
representation of the subsetting expression for \code{dt}.}

\item{transformation_arguments}{[\code{list()}]\cr
NULL if no transformation. Otherwise list of arguments to provide to the
corresponding \code{transformation} function(s). Should be the same length and
use the same naming as \code{transformation}.}
}
\value{
invisibly return \code{dt} with transformed values.
}
\description{
Helper function to transform values in a data.table
}
