% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popReconstruct_fit.R
\name{validate_popReconstruct_inputs}
\alias{validate_popReconstruct_inputs}
\title{Helper function to validate initial input data.tables}
\usage{
validate_popReconstruct_inputs(inputs, settings, detailed_settings, value_col)
}
\arguments{
\item{inputs}{[\code{list()}]\cr
[\code{data.table()}] for each ccmpp input. Requires 'srb', 'asfr', 'baseline',
and 'survival'; and migration estimates provided as just 'net_migration' or
both 'immigration' and 'emigration'. See \strong{Section: Inputs} for more
information on each of the required inputs.}

\item{settings}{[\code{list()}]\cr
Named list of settings for running the popReconstruct model with. The
required settings are the same as those required for \code{\link[demCore:ccmpp]{demCore::ccmpp()}},
see \strong{Section: Settings} for these. The optional settings specific to
popReconstruct are described in
\strong{Section: Optional popReconstruct Settings}}

\item{detailed_settings}{[\code{list()}]\cr
Detailed settings for each ccmpp input and 'population'.}

\item{value_col}{[\code{character(1)}]\cr
Name of the column containing the value of interest in each of the
\code{inputs}. Default is 'value'.}
}
\description{
Assert that each of the inputs once transformed is a finite
value greater than negative Infinity and less than positive Infinity.
}
